Backend1/requirements.txt
fastapi
uvicorn[standard]
asyncpg
httpx[http2]>=0.27
python-dotenv
google-generativeai
orjson
requests
pydantic
asyncpg